requirements.txt
-i https://pypi.org/simple
asgiref==3.9.1; python_version >= '3.9'
certifi==2025.8.3; python_version >= '3.7'
charset-normalizer==3.4.3; python_version >= '3.7'
coreapi==2.3.3
coreschema==0.0.4
django==5.2.5
djangorestframework==3.16.1
drf-yasg==1.21.10
idna==3.10; python_version >= '3.6'
inflection==0.5.1; python_version >= '3.5'
itypes==1.2.0
jinja2==3.1.6; python_version >= '3.7'
markupsafe==3.0.2; python_version >= '3.9'
packaging==25.0; python_version >= '3.8'
psycopg2==2.9.10
python-decouple==3.8
pytz==2025.2
pyyaml==6.0.2; python_version >= '3.8'
requests==2.32.5; python_version >= '3.9'
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3; python_version >= '3.8'
tzdata==2025.2; sys_platform == 'win32'
uritemplate==4.2.0; python_version >= '3.9'
urllib3==2.5.0; python_version >= '3.9'
